refactor: split utils folder
