+++ src/utils/constants/people-unknown.const.ts
@@ -1,0 +1,3 @@
+import { PeopleUnknown } from '../interfaces-types-enums/people.interface';
+
+export const peopleUnknown: PeopleUnknown = { detail: 'Not found' };
