+++ src/utils/constants/text.const.ts
