+++ src/utils/index.ts
@@ -1,12 +1,12 @@
-export * from './key-value-pair.type';
-export * from './loading-state.enum';
-export * from './people.interface';
-export * from './search-result.interface';
-export * from './path-value.enum';
-export * from './get-id-from-url';
-export * from './get-people-formatted';
-export * from './people-formatted.interface';
-export * from './get-people-formatted';
-export * from './get-id-from-url';
-export * from './text.const';
-export * from './people-unknown.const';
+export * from './interfaces-types-enums/key-value-pair.type';
+export * from './interfaces-types-enums/loading-state.enum';
+export * from './interfaces-types-enums/people.interface';
+export * from './interfaces-types-enums/search-result.interface';
+export * from './interfaces-types-enums/path-value.enum';
+export * from './functions/get-id-from-url';
+export * from './functions/get-people-formatted';
+export * from './interfaces-types-enums/people-formatted.interface';
+export * from './functions/get-people-formatted';
+export * from './functions/get-id-from-url';
+export * from './constants/text.const';
+export * from './constants/people-unknown.const';
