+++ src/utils/interfaces-types-enums/loading-state.enum.ts
