+++ src/utils/interfaces-types-enums/search-result.interface.ts
