--- src/utils/people-unknown.const.ts
@@ -1,3 +1,0 @@
-import { PeopleUnknown } from './people.interface';
-
-export const peopleUnknown: PeopleUnknown = { detail: 'Not found' };
